Not sure this lab will get done
This will be my basis for lab 7 a and b. Followed zales example but
built it with ints in mind when I have Strings to worry about. I will
probs just add the string value to the nodes. Also for the weighted
graph I will need to add an edge class.